extract renderStatus into a ComputedStatus componennt
