+++ src/app/js/admin/precomputed/PrecomputedForm.tsx
@@ -11,7 +11,6 @@ import PrecomputedPreview from './PrecomputedPreview';
 import PrecomputedFormLogsDialogComponent from './PrecomputedLogsDialog';
 import PrecomputedFormDataDialogComponent from './PrecomputedDataDialog';
 import PlayArrowIcon from '@mui/icons-material/PlayArrow';
-import PropTypes from 'prop-types';
 import SourceValueFromColumns from './SourceValueFromColumns';
 import { useForm } from 'react-hook-form';
 import { TextField } from '../../reactHookFormFields/TextField';
@@ -23,7 +22,7 @@ import { compose } from 'recompose';
 import { fromPrecomputed, fromParsing } from '../selectors';
 import { ListAlt as ListAltIcon } from '@mui/icons-material';
 import { withRouter } from 'react-router';
-import { Box, Button, Chip, Typography, type ButtonProps } from '@mui/material';
+import { Box, Button, Typography, type ButtonProps } from '@mui/material';
 import {
     createPrecomputed,
     getPreviewPrecomputed,
@@ -36,164 +35,16 @@ import {
     FINISHED,
     IN_PROGRESS,
     PENDING,
-    ERROR,
-    CANCELED,
-    PAUSED,
     ON_HOLD,
 } from '../../../../common/taskStatus';
 import { io } from 'socket.io-client';
 import CancelButton from '../../lib/components/CancelButton';
 import { DEFAULT_TENANT } from '../../../../common/tools/tenantTools';
-import getLocale from '../../../../common/getLocale';
 import { useTranslate } from '../../i18n/I18NContext';
+import { PrecomputedStatus } from './PrecomputedStatus';
 
 const required = (text: string) => (value: unknown) =>
     value && !(value instanceof Array && value.length === 0) ? undefined : text;
-
-function getDisplayTimeStartedAt(startedAt: string) {
-    if (!startedAt) {
-        return;
-    }
-
-    const now = new Date();
-    const startedAtDate = new Date(startedAt);
-    const diff = now.getTime() - startedAtDate.getTime();
-
-    const diffInMinutes = Math.floor(diff / 60000);
-    const diffInHours = Math.floor(diffInMinutes / 60);
-    const diffInDays = Math.floor(diffInHours / 24);
-
-    const relativeTime = new Intl.RelativeTimeFormat(getLocale(), {
-        numeric: 'auto',
-    });
-    let timeSinceStarted = '';
-
-    if (diffInHours < 1) {
-        timeSinceStarted = relativeTime.format(-diffInMinutes, 'minute');
-    } else if (diffInDays < 1) {
-        timeSinceStarted = relativeTime.format(-diffInHours, 'hour');
-    } else {
-        timeSinceStarted = relativeTime.format(-diffInDays, 'day');
-    }
-    return timeSinceStarted;
-}
-
-export const StatusChip = ({
-    label,
-    color,
-    startedAt,
-    sx,
-}: {
-    label: string;
-    color?:
-        | 'default'
-        | 'primary'
-        | 'secondary'
-        | 'error'
-        | 'info'
-        | 'success'
-        | 'warning';
-    startedAt?: string | null;
-    sx?: object;
-}) => {
-    const [spentTime, setSpentTime] = useState(
-        startedAt ? getDisplayTimeStartedAt(startedAt) : '',
-    );
-    useEffect(() => {
-        if (!startedAt) return;
-        const interval = setInterval(() => {
-            setSpentTime(getDisplayTimeStartedAt(startedAt));
-        }, 59000);
-        return () => clearInterval(interval);
-    }, [startedAt]);
-    const finalLabel = `${label}${startedAt ? ` (${spentTime})` : ''}`;
-    return <Chip component="span" label={finalLabel} color={color} sx={sx} />;
-};
-
-StatusChip.propTypes = {
-    color: PropTypes.string,
-    label: PropTypes.string.isRequired,
-    startedAt: PropTypes.string,
-    sx: PropTypes.object,
-};
-
-export const renderStatus = (
-    status: string,
-    translate: (key: string) => string,
-    startedAt: string | null = null,
-) => {
-    if (status === PENDING) {
-        return (
-            <StatusChip
-                label={translate('precomputed_status_pending')}
-                color="warning"
-                startedAt={startedAt}
-            />
-        );
-    }
-    if (status === IN_PROGRESS) {
-        return (
-            <StatusChip
-                label={translate('precomputed_status_running')}
-                color="info"
-                startedAt={startedAt}
-            />
-        );
-    }
-
-    if (status === PAUSED) {
-        return (
-            <StatusChip
-                label={translate('precomputed_status_paused')}
-                color="info"
-            />
-        );
-    }
-
-    if (status === FINISHED) {
-        return (
-            <StatusChip
-                label={translate('precomputed_status_done')}
-                color="success"
-            />
-        );
-    }
-
-    if (status === ERROR) {
-        return (
-            <StatusChip
-                label={translate('precomputed_status_error')}
-                color="error"
-            />
-        );
-    }
-
-    if (status === CANCELED) {
-        return (
-            <StatusChip
-                label={translate('precomputed_status_canceled')}
-                color="warning"
-            />
-        );
-    }
-
-    if (status === ON_HOLD) {
-        return (
-            <StatusChip
-                label={translate('precomputed_status_hold')}
-                sx={{ backgroundColor: '#539CE1', color: '#fff' }}
-                startedAt={startedAt}
-            />
-        );
-    }
-
-    return (
-        <StatusChip
-            label={translate('precomputed_status_not_started')}
-            sx={{ backgroundColor: 'neutral' }}
-        />
-    );
-};
 
 export const RunButton = ({
     handleLaunchPrecomputed,
@@ -490,11 +341,11 @@ export const PrecomputedForm = ({
                             >
                                 <Typography>
                                     {translate('precomputed_status')} : &nbsp;
-                                    {renderStatus(
-                                        precomputedStatus,
-                                        translate,
-                                        initialValues?.startedAt,
-                                    )}
+                                    <PrecomputedStatus
+                                        status={precomputedStatus}
+                                        translate={translate}
+                                        startedAt={initialValues?.startedAt}
+                                    />
                                 </Typography>
                                 <Box>
                                     <Button
