+++ src/app/js/admin/precomputed/PrecomputedList.tsx
@@ -13,12 +13,13 @@ import {
 } from '@mui/x-data-grid';
 import { Link } from 'react-router-dom';
 import { useHistory } from 'react-router';
-import { renderStatus, RunButton } from './PrecomputedForm';
+import { RunButton } from './PrecomputedForm';
 import { fromPrecomputed } from '../selectors';
 import { launchPrecomputed } from '.';
 import { IN_PROGRESS, FINISHED, ON_HOLD } from '../../../../common/taskStatus';
 import { toast } from '../../../../common/tools/toast';
 import { useTranslate } from '../../i18n/I18NContext';
+import { PrecomputedStatus } from './PrecomputedStatus';
 
 export const PrecomputedList = ({
     // @ts-expect-error TS7031
@@ -105,12 +106,13 @@ export const PrecomputedList = ({
                         field: 'status',
                         headerName: translate('precomputed_status'),
                         flex: 3,
-                        renderCell: (params) =>
-                            renderStatus(
-                                params.row.status,
-                                translate,
-                                params.row.startedAt,
-                            ),
+                        renderCell: (params) => (
+                            <PrecomputedStatus
+                                status={params.row.status}
+                                translate={translate}
+                                startedAt={params.row.startedAt}
+                            />
+                        ),
                     },
                     {
                         field: 'run',
@@ -159,10 +161,8 @@ PrecomputedList.propTypes = {
 const mapStateToProps = (state) => ({
     precomputedList: state.precomputed.precomputed,
     isPrecomputedRunning: !!fromPrecomputed
-        // @ts-expect-error TS2339
         .precomputed(state)
         .find(
-            // @ts-expect-error TS7006
             (precomputedData) =>
                 precomputedData.status === IN_PROGRESS ||
                 precomputedData.status === ON_HOLD,
